+++ resourcet-extra/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version 0.0.2
 
-* Add `runResourceT`, `asyncRestore`.
+* Add `runResourceT`, `asyncRestore`, `releaseType`, `unprotectType`.
 
 
 # Version 0.0.1